Fix for https://github.com/crkn-rcdr/CIHM-TDR/issues/7
Rather than trying to upload 3 times and only checking validation once, we now try 5 times and couple a validation with each upload attempt -- this means that the validation can fail as well and retry.

I also added a 30 second delay before the retry, given most of the errors so far have been because of pending container updates.